Move solidus initial temperature to test
